fix: move processor to subfolder
